requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.22
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
